requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.23.0
anthropic>=0.39.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
ddgs>=7.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
